fixed issues caused by divergent levelManager
