+++ Assets/BaseScene.unity
@@ -262,87 +262,77 @@ Transform:
   m_Father: {fileID: 1805071817}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!1001 &1678443511
-Prefab:
-  m_ObjectHideFlags: 0
-  serializedVersion: 2
-  m_Modification:
-    m_TransformParent: {fileID: 0}
-    m_Modifications:
-    - target: {fileID: 114508715870727834, guid: 1067e80f8d8275241a2b586aee84bfbe,
-        type: 2}
-      propertyPath: Enemies.Array.size
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 4334807874380840, guid: 1067e80f8d8275241a2b586aee84bfbe, type: 2}
-      propertyPath: m_LocalPosition.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 4334807874380840, guid: 1067e80f8d8275241a2b586aee84bfbe, type: 2}
-      propertyPath: m_LocalPosition.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 4334807874380840, guid: 1067e80f8d8275241a2b586aee84bfbe, type: 2}
-      propertyPath: m_LocalPosition.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 4334807874380840, guid: 1067e80f8d8275241a2b586aee84bfbe, type: 2}
-      propertyPath: m_LocalRotation.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 4334807874380840, guid: 1067e80f8d8275241a2b586aee84bfbe, type: 2}
-      propertyPath: m_LocalRotation.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 4334807874380840, guid: 1067e80f8d8275241a2b586aee84bfbe, type: 2}
-      propertyPath: m_LocalRotation.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 4334807874380840, guid: 1067e80f8d8275241a2b586aee84bfbe, type: 2}
-      propertyPath: m_LocalRotation.w
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 4334807874380840, guid: 1067e80f8d8275241a2b586aee84bfbe, type: 2}
-      propertyPath: m_RootOrder
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 114508715870727834, guid: 1067e80f8d8275241a2b586aee84bfbe,
-        type: 2}
-      propertyPath: player
-      value: 
-      objectReference: {fileID: 1780647233420238, guid: 84d5fa99bfd488747bf2969b8fa96a23,
-        type: 2}
-    - target: {fileID: 114590791460387946, guid: 1067e80f8d8275241a2b586aee84bfbe,
-        type: 2}
-      propertyPath: playersTurn
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 114590791460387946, guid: 1067e80f8d8275241a2b586aee84bfbe,
-        type: 2}
-      propertyPath: playerFoodPoints
-      value: 100
-      objectReference: {fileID: 0}
-    - target: {fileID: 114508715870727834, guid: 1067e80f8d8275241a2b586aee84bfbe,
-        type: 2}
-      propertyPath: Enemies.Array.data[0]
-      value: 
-      objectReference: {fileID: 1855614962614306, guid: 68e77259e4be873439f109b9a7292e84,
-        type: 2}
-    - target: {fileID: 114508715870727834, guid: 1067e80f8d8275241a2b586aee84bfbe,
-        type: 2}
-      propertyPath: columns
-      value: 15
-      objectReference: {fileID: 0}
-    - target: {fileID: 114508715870727834, guid: 1067e80f8d8275241a2b586aee84bfbe,
-        type: 2}
-      propertyPath: rows
-      value: 15
-      objectReference: {fileID: 0}
-    m_RemovedComponents: []
-  m_ParentPrefab: {fileID: 100100000, guid: 1067e80f8d8275241a2b586aee84bfbe, type: 2}
-  m_IsPrefabParent: 0
 --- !u!4 &1805071817 stripped
 Transform:
   m_PrefabParentObject: {fileID: 4807465879003910, guid: 84d5fa99bfd488747bf2969b8fa96a23,
     type: 2}
   m_PrefabInternal: {fileID: 723018084}
+--- !u!1 &1978801240
+GameObject:
+  m_ObjectHideFlags: 0
+  m_PrefabParentObject: {fileID: 1412084579755790, guid: 1067e80f8d8275241a2b586aee84bfbe,
+    type: 2}
+  m_PrefabInternal: {fileID: 0}
+  serializedVersion: 5
+  m_Component:
+  - component: {fileID: 1978801243}
+  - component: {fileID: 1978801242}
+  - component: {fileID: 1978801241}
+  m_Layer: 0
+  m_Name: GameManager
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!114 &1978801241
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_PrefabParentObject: {fileID: 114508715870727834, guid: 1067e80f8d8275241a2b586aee84bfbe,
+    type: 2}
+  m_PrefabInternal: {fileID: 0}
+  m_GameObject: {fileID: 1978801240}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 359764853adfa504ab6729475996d35e, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  Columns: 8
+  Rows: 8
+  Walls:
+  - {fileID: 1126637172613874, guid: da06bfd96e18e6141bccb22e8f52e66e, type: 2}
+  Floors:
+  - {fileID: 1107012629488888, guid: 4b6ecfed334523d4b80d9a04140b840b, type: 2}
+  Enemies: []
+  Loot: []
+  Doors: []
+  Obstacles: []
+--- !u!114 &1978801242
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_PrefabParentObject: {fileID: 114590791460387946, guid: 1067e80f8d8275241a2b586aee84bfbe,
+    type: 2}
+  m_PrefabInternal: {fileID: 0}
+  m_GameObject: {fileID: 1978801240}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 86b9d09d417ff6943ba1fb04e69ead27, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  maxLevel: 15
+  playersTurn: 1
+  playerHealth: 0
+--- !u!4 &1978801243
+Transform:
+  m_ObjectHideFlags: 0
+  m_PrefabParentObject: {fileID: 4334807874380840, guid: 1067e80f8d8275241a2b586aee84bfbe,
+    type: 2}
+  m_PrefabInternal: {fileID: 0}
+  m_GameObject: {fileID: 1978801240}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 0, y: 0, z: 0}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_Children: []
+  m_Father: {fileID: 0}
+  m_RootOrder: 0
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
